Extract code related only to controlling Blinkenlights out of the script.js into a separate blinken.js. This is necessary to be able to reuse the blinkenlight js methods from both the main page and the blinkmoji.
